css/css-lists/parsing/counter-reset-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Lists: getComputedStyle().counterReset</title>
<link rel="help" href="https://drafts.csswg.org/css-lists-3/#propdef-counter-reset">
<meta name="assert" content="counter-reset computed value is the keyword none or a list, each item an identifier paired with an integer.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value('counter-reset', 'none');

test_computed_value('counter-reset', 'section -1');

// Not yet tested:
// Serialization with distinct and repeated counter names
// https://github.com/w3c/csswg-drafts/issues/4216
</script>
</body>
</html>
